pyproject.toml
[project]
name = "py-simple-auth-api"
version = "0.1.0"
description = ""
authors = [
    {name = "romaopatrick"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "lagom (>=2.7.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
